feat: Enhance SocialSecurity component with date formatting and debugging utilities
- Added a debug utility function to log date formats for payroll periods.
- Implemented consistent date formatting for payroll entries when fetching from the backend.
- Improved error handling and logging for date parsing and formatting.
- Created a new debugUtils module for API response logging and ID extraction.
- Introduced functions to verify if a payroll was created successfully and to extract payroll IDs from various response formats.

+++ BACKEND_PAYROLL/src/main/java/co/edu/unbosque/backend_payroll/controller/LiquidacionCesantiasController.java
@@ -29,7 +29,7 @@ public class LiquidacionCesantiasController {
     }
 
     @GetMapping("/disponible/{contratoId}")
-    public ResponseEntity<Map<String, Object>> consultarCesantiasDisponibles(@PathVariable Integer contratoId) {
+    public ResponseEntity<Map<String, Object>> consultarCesantiasDisponibles(@PathVariable Short contratoId) {
         try {
             Map<String, Object> cesantias = service.consultarCesantiasDisponibles(contratoId);
             return ResponseEntity.ok(cesantias);
